fix: link pet-setting-illustrated_guide,nullref#20
